miorai_backend/requirements.txt
Django>=5.2.1
djangorestframework>=3.16.0
django-rest-knox>=4.2.0
django-cors-headers>=4.7.0
mysqlclient>=2.2.0
Pillow>=10.0.0
django-ratelimit>=4.1.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
# ML Libraries
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0 